Revert "Move PASSWORD_PLACEHOLDER to hqwebapp"
This reverts commit 68f4670d8d765157df39199c59efafddb7f72271.
