+++ corehq/apps/dump_reload/sql/serialization.py
@@ -3,7 +3,7 @@ from copy import copy
 
 from django.core.serializers.json import Serializer as JsonSerializer
 
-from corehq.apps.hqwebapp.const import PASSWORD_PLACEHOLDER
+from corehq.motech.const import PASSWORD_PLACEHOLDER
 from corehq.util.json import CommCareJSONEncoder
