--- corehq/apps/hqwebapp/const.py
@@ -1,1 +1,0 @@
-PASSWORD_PLACEHOLDER = '*' * 16
