+++ corehq/motech/openmrs/views.py
@@ -15,9 +15,9 @@ from dimagi.utils.web import json_response
 from corehq import toggles
 from corehq.apps.domain.decorators import login_and_domain_required
 from corehq.apps.domain.views.settings import BaseProjectSettingsView
-from corehq.apps.hqwebapp.const import PASSWORD_PLACEHOLDER
 from corehq.apps.users.decorators import require_permission
 from corehq.apps.users.models import HqPermissions
+from corehq.motech.const import PASSWORD_PLACEHOLDER
 from corehq.motech.openmrs.dbaccessors import get_openmrs_importers_by_domain
 from corehq.motech.openmrs.forms import (
     OpenmrsConfigForm,
