+++ corehq/motech/tests/test_models.py
@@ -8,8 +8,7 @@ import requests
 from unittest.mock import Mock, patch
 
 from corehq.apps.hqwebapp.templatetags.hq_shared_tags import pp_json
-from corehq.apps.hqwebapp.const import PASSWORD_PLACEHOLDER
-from corehq.motech.const import ALGO_AES
+from corehq.motech.const import ALGO_AES, PASSWORD_PLACEHOLDER
 from corehq.motech.models import (
     ConnectionSettings,
     RequestLog,
